+++ docs/CHANGELOG.rst
@@ -26,6 +26,7 @@ Added
 -----
 - Cap process memory consumption at 16GB
 - Add support for Python 3.9
+- Add ``wgs-preprocess`` process
 
 
 ===================